Adds image download functionality for cards
Implements the ability to download images for cards
retrieved from the Lorcast API.

Adds a new `download_image` method to the `LorcastAPI` class
to handle image downloading and saving.

Extends the `InkcollectorCLI` to include a `--get-images` flag
to the `get-cards` command. When specified, the script will
download the images for each card in the retrieved set,
saving them to a new "images" directory, organized by datasource
and set ID.

Adds "images/" to the .gitignore file.

Refactors the output directory setup.